+++ x/ibc/07-tendermint/keeper/misbehaviour_test.go
@@ -1,4 +1,4 @@
-package tendermint_test
+package keeper_test
 
 import (
 	"bytes"
