+++ x/ibc/07-tendermint/keeper/update.go
@@ -1,4 +1,4 @@
-package tendermint
+package keeper
 
 import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
