+++ x/ibc/07-tendermint/types/codec.go
@@ -1,4 +1,4 @@
-package tendermint
+package types
 
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
