+++ x/ibc/07-tendermint/types/consensus_state.go
@@ -1,4 +1,4 @@
-package tendermint
+package types
 
 import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
