+++ x/ibc/07-tendermint/types/evidence.go
@@ -1,4 +1,4 @@
-package tendermint
+package types
 
 import (
 	"math"
