+++ x/ibc/07-tendermint/types/evidence_test.go
@@ -1,4 +1,4 @@
-package tendermint_test
+package types_test
 
 import (
 	"bytes"
