+++ x/ibc/07-tendermint/types/header.go
@@ -1,4 +1,4 @@
-package tendermint
+package types
 
 import (
 	abci "github.com/tendermint/tendermint/abci/types"
